+++ backend/sql/schema/003_plan_categories.sql
@@ -3,8 +3,8 @@ CREATE TABLE plan_categories(
     id UUID PRIMARY KEY,
     plan_id UUID NOT NULL REFERENCES plans(id),
     name TEXT NOT NULL,
-    deposit INT NOT NULL,
-    withdrawal INT NOT NULL,
+    deposit BIGINT NOT NULL,
+    withdrawal BIGINT NOT NULL,
     created_at TIMESTAMP NOT NULL,
     updated_at TIMESTAMP NOT NULL,
     UNIQUE(plan_id, name)
